move source loading and serialization into transform module
