+++ libs/datamodel/core/src/common/provider_names.rs
@@ -1,0 +1,5 @@
+pub const SQLITE_SOURCE_NAME: &str = "sqlite";
+pub const POSTGRES_SOURCE_NAME: &str = "postgresql";
+pub const POSTGRES_SOURCE_NAME_HEROKU: &str = "postgres";
+pub const MYSQL_SOURCE_NAME: &str = "mysql";
+pub const MSSQL_SOURCE_NAME: &str = "sqlserver";
