+++ libs/datamodel/core/src/configuration/datasource.rs
@@ -1,5 +1,5 @@
-use crate::StringFromEnvVar;
 use datamodel_connector::Connector;
+use serde::Serialize;
 
 /// a `datasource` from the prisma schema.
 pub struct Datasource {
@@ -20,4 +20,12 @@ impl Datasource {
     pub fn url(&self) -> &StringFromEnvVar {
         &self.url
     }
+}
+
+#[serde(rename_all = "camelCase")]
+#[derive(Clone, Debug, Serialize, PartialEq)]
+pub struct StringFromEnvVar {
+    /// contains the name of env var if the value was read from one
+    pub from_env_var: Option<String>,
+    pub value: String,
 }
