+++ libs/datamodel/core/src/configuration/mod.rs
@@ -1,20 +1,10 @@
+mod datasource;
 mod generator;
-mod source;
 
+pub use datasource::*;
 pub use generator::*;
-pub use source::*;
-
-use serde::Serialize;
 
 pub struct Configuration {
     pub generators: Vec<Generator>,
     pub datasources: Vec<Datasource>,
-}
-
-#[serde(rename_all = "camelCase")]
-#[derive(Clone, Debug, Serialize, PartialEq)]
-pub struct StringFromEnvVar {
-    /// contains the name of env var if the value was read from one
-    pub from_env_var: Option<String>,
-    pub value: String,
 }
