--- libs/datamodel/core/src/configuration/source/mod.rs
@@ -1,14 +1,0 @@
-mod builtin_datasource_providers;
-mod datasource;
-mod datasource_provider;
-mod loader;
-mod serializer;
-
-pub use builtin_datasource_providers::{
-    MsSqlDatasourceProvider, MySqlDatasourceProvider, PostgresDatasourceProvider, SqliteDatasourceProvider,
-    MSSQL_SOURCE_NAME, MYSQL_SOURCE_NAME, POSTGRES_SOURCE_NAME, SQLITE_SOURCE_NAME,
-};
-pub use datasource::Datasource;
-pub use datasource_provider::DatasourceProvider;
-pub use loader::SourceLoader;
-pub use serializer::SourceSerializer;
