+++ libs/datamodel/core/src/transform/ast_to_dml/mod.rs
@@ -1,4 +1,7 @@
+mod builtin_datasource_providers;
 mod common;
+mod datasource_loader;
+mod datasource_provider;
 mod generator_loader;
 mod invalid_model_names;
 mod lift;
@@ -7,8 +10,10 @@ mod standardise;
 mod validate;
 mod validation_pipeline;
 
-pub use generator_loader::GeneratorLoader;
 use lift::*;
 use standardise::*;
 use validate::*;
+
+pub use datasource_loader::DatasourceLoader;
+pub use generator_loader::GeneratorLoader;
 pub use validation_pipeline::ValidationPipeline;
