+++ libs/datamodel/core/src/transform/dml_to_ast/mod.rs
@@ -1,5 +1,7 @@
+mod datasource_serializer;
 mod generator_serializer;
 mod lower;
 
+pub use datasource_serializer::DatasourceSerializer;
 pub use generator_serializer::GeneratorSerializer;
 pub use lower::LowerDmlToAst;
